Solve problem with FLX link IP core update
It happens that the XCI file stores the original path of when the IP
Core was created. The FLX link IP Core was created in a standalone
environment for testing purposes and then include in the main project
later. To fix this issue, I enabled its core container and then
disabled it, causing the IP Core XCI to be regenerated, now with
the correct hardcoded path.